pyproject.toml
[project]
name = "diabimmune-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "matplotlib>=3.10.7",
    "pyyaml>=6.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
]
